+++ src/middleware/mod.rs
@@ -2,7 +2,6 @@
 mod compress;
 pub use self::compress::{BodyEncoding, Compress};
 
-pub mod cors;
 mod defaultheaders;
 pub mod errhandlers;
 mod logger;
@@ -11,3 +10,6 @@ mod normalize;
 pub use self::defaultheaders::DefaultHeaders;
 pub use self::logger::Logger;
 pub use self::normalize::NormalizePath;
+
+//
+// use actix_cors as cors;