+++ actix-cors/src/lib.rs
@@ -7,7 +7,7 @@
 //! # Example
 //!
 //! ```rust
-//! use actix_web::middleware::cors::Cors;
+//! use actix_cors::Cors;
 //! use actix_web::{http, web, App, HttpRequest, HttpResponse, HttpServer};
 //!
 //! fn index(req: HttpRequest) -> &'static str {
@@ -42,16 +42,14 @@ use std::iter::FromIterator;
 use std::rc::Rc;
 
 use actix_service::{IntoTransform, Service, Transform};
+use actix_web::dev::{RequestHead, ServiceRequest, ServiceResponse};
+use actix_web::error::{Error, ResponseError, Result};
+use actix_web::http::header::{self, HeaderName, HeaderValue};
+use actix_web::http::{self, HttpTryFrom, Method, StatusCode, Uri};
+use actix_web::HttpResponse;
 use derive_more::Display;
 use futures::future::{ok, Either, Future, FutureResult};
 use futures::Poll;
-
-use crate::dev::RequestHead;
-use crate::error::{Error, ResponseError, Result};
-use crate::http::header::{self, HeaderName, HeaderValue};
-use crate::http::{self, HttpTryFrom, Method, StatusCode, Uri};
-use crate::service::{ServiceRequest, ServiceResponse};
-use crate::HttpResponse;
 
 /// A set of errors that can occur during processing CORS
 #[derive(Debug, Display)]
@@ -152,11 +150,11 @@ impl<T> AllOrSome<T> {
 /// # Example
 ///
 /// ```rust
+/// use actix_cors::Cors;
 /// use actix_web::http::header;
-/// use actix_web::middleware::cors;
 ///
 /// # fn main() {
-/// let cors = cors::Cors::new()
+/// let cors = Cors::new()
 ///     .allowed_origin("https://www.rust-lang.org/")
 ///     .allowed_methods(vec!["GET", "POST"])
 ///     .allowed_headers(vec![header::AUTHORIZATION, header::ACCEPT])
@@ -806,9 +804,9 @@ where
 #[cfg(test)]
 mod tests {
     use actix_service::{IntoService, Transform};
+    use actix_web::test::{self, block_on, TestRequest};
 
     use super::*;
-    use crate::test::{self, block_on, TestRequest};
 
     impl Cors {
         fn finish<F, S, B>(self, srv: F) -> CorsMiddleware<S>