+++ src/rmap.rs
@@ -38,7 +38,8 @@ impl ResourceMap {
     pub(crate) fn finish(&self, current: Rc<ResourceMap>) {
         for (_, nested) in &self.patterns {
             if let Some(ref nested) = nested {
-                *nested.parent.borrow_mut() = Some(current.clone())
+                *nested.parent.borrow_mut() = Some(current.clone());
+                nested.finish(nested.clone());
             }
         }
     }